Move to `create_app.py`
